refactor: split server.py (3536→2361 lines) into focused route modules
Extract endpoint handlers from the monolithic server.py into 5 APIRouter modules:
- routes_decide.py: /v1/decide, /v1/replay, /v1/fuji/validate
- routes_memory.py: /v1/memory/* CRUD endpoints
- routes_trust.py: /v1/trust/*, /v1/trustlog/* endpoints
- routes_governance.py: /v1/governance/* with RBAC
- routes_system.py: health, status, metrics, SSE, compliance, reports, system halt

server.py retains middleware, auth, lazy loading, and re-exports all
attributes for full backward compatibility with existing test monkeypatching.
