+++ veritas_os/api/server.py
@@ -2114,59 +2114,6 @@ def _is_direct_fuji_api_enabled() -> bool:
     return normalized_flag in truthy_values
 
 
-@app.get("/")
-def root() -> Dict[str, Any]:
-    return {"ok": True, "service": "veritas-api", "server_time": utc_now_iso_z()}
-
-
-@app.get("/health")
-@app.get("/v1/health")
-def health() -> Dict[str, Any]:
-    try:
-        pipeline_ok = get_decision_pipeline() is not None
-        memory_ok = get_memory_store() is not None
-        all_ok = pipeline_ok and memory_ok
-        result: Dict[str, Any] = {
-            "ok": all_ok,
-            "uptime": int(time.time() - START_TS),
-            "checks": {
-                "pipeline": "ok" if pipeline_ok else "unavailable",
-                "memory": "ok" if memory_ok else "unavailable",
-            },
-        }
-        return result
-    except Exception as e:
-        logger.error("[Health] check failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "health check failed"},
-        )
-
-
-@app.get("/status")
-@app.get("/v1/status")
-@app.get("/api/status")
-def status() -> Dict[str, Any]:
-    expected = (_get_expected_api_key() or "").strip()
-    result = {
-        "ok": True,
-        "version": "veritas-api 1.0.3",
-        "uptime": int(time.time() - START_TS),
-        "server_time": utc_now_iso_z(),
-        "pipeline_ok": get_decision_pipeline() is not None,
-        "api_key_configured": bool(expected),
-    }
-    # ★ M-13 修正: 内部エラー詳細はデバッグモード時のみ公開
-    # 本番環境では実装の詳細が漏洩するのを防止
-    if _is_debug_mode():
-        result["cfg_error"] = _cfg_state.err
-        result["pipeline_error"] = _pipeline_state.err
-    else:
-        result["cfg_error"] = bool(_cfg_state.err)
-        result["pipeline_error"] = bool(_pipeline_state.err)
-    return result
-
-
 # ==============================
 # Trust log helpers（server 側でも軽く読めるように）
 # ==============================
@@ -2327,1210 +2274,88 @@ def write_shadow_decide(
 
 
 # ==============================
-# main: /v1/decide
+# Route modules (APIRouter)
 # ==============================
+# All endpoint handlers have been extracted into focused route modules.
+# Each router is included below with shared auth/rate-limit dependencies.
+# server.py remains the single FastAPI app entry point and re-exports
+# all attributes that tests monkeypatch.
 
-@app.post(
-    "/v1/decide",
-    response_model=DecideResponse,
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)],
+from veritas_os.api.routes_decide import router as _decide_router  # noqa: E402
+from veritas_os.api.routes_decide import (  # noqa: E402,F401 -- backward compat
+    _call_fuji,
+    decide,
+    replay_endpoint,
+    replay_decision_endpoint,
+    fuji_validate,
 )
-async def decide(req: DecideRequest, request: Request):
-    p = get_decision_pipeline()
-    if p is None:
-        _log_decide_failure("decision_pipeline unavailable", _pipeline_state.err)
-        _publish_event("decide.completed", {"ok": False, "error": DECIDE_GENERIC_ERROR})
-        return JSONResponse(
-            status_code=503,
-            content={
-                "ok": False,
-                "error": DECIDE_GENERIC_ERROR,
-                "detail": DECIDE_GENERIC_ERROR,
-                "trust_log": None,  # ★互換
-            },
-        )
-
-    try:
-        payload = await p.run_decide_pipeline(req=req, request=request)
-    except Exception as e:
-        failure_category = _classify_decide_failure(e)
-        _log_decide_failure("decision_pipeline execution failed", e)
-        _publish_event(
-            "decide.completed",
-            {
-                "ok": False,
-                "error": DECIDE_GENERIC_ERROR,
-                "failure_category": failure_category,
-            },
-        )
-        return JSONResponse(
-            status_code=503,
-            content={
-                "ok": False,
-                "error": DECIDE_GENERIC_ERROR,
-                "detail": DECIDE_GENERIC_ERROR,
-                "failure_category": failure_category,
-                "trust_log": None,  # ★互換
-            },
-        )
-
-    if isinstance(payload, dict):
-        resolve_dynamic_steps(payload)
-        _publish_event(
-            "trustlog.debate",
-            {
-                "request_id": payload.get("request_id"),
-                "summary": _stage_summary(
-                    payload.get("debate"),
-                    "debate stage completed",
-                ),
-            },
-        )
-        _publish_event(
-            "trustlog.critique",
-            {
-                "request_id": payload.get("request_id"),
-                "summary": _stage_summary(
-                    payload.get("critique"),
-                    "critique stage completed",
-                ),
-            },
-        )
-
-    # ★ response_model を効かせつつ壊れpayloadでも落ちないように最終整形
-    coerced = _coerce_decide_payload(payload, seed=getattr(req, "query", "") or "")
-    try:
-        coerced = enforce_compliance_stop(coerced)
-    except ComplianceStopException as stop:
-        _publish_event(
-            "compliance.pending_review",
-            {
-                "request_id": stop.payload.get("request_id"),
-                "status": stop.payload.get("status"),
-            },
-        )
-        return JSONResponse(status_code=200, content=stop.payload)
-
-    # DecideResponse 側に "落ちない" バリデータがある前提で model_validate
-    try:
-        _publish_event(
-            "decide.completed",
-            {
-                "ok": bool(coerced.get("ok", True)),
-                "request_id": coerced.get("request_id"),
-                "decision": coerced.get("decision"),
-            },
-        )
-        fuji_payload = coerced.get("fuji") or {}
-        if str(fuji_payload.get("status", "")).lower() in {"reject", "rejected", DECISION_REJECTED}:
-            _publish_event(
-                "fuji.rejected",
-                {
-                    "request_id": coerced.get("request_id"),
-                    "status": fuji_payload.get("status"),
-                    "reasons": fuji_payload.get("reasons", []),
-                },
-            )
-        return DecideResponse.model_validate(coerced)
-    except Exception as e:
-        # 最後の保険：型検証で落ちても 503 ではなく JSON にして "落ちない" を優先
-        # ok=False でクライアントが正常レスポンスと区別できるようにする
-        # ★ セキュリティ修正: 内部エラー詳細はデバッグモード時のみ公開
-        logger.error("DecideResponse validation failed: %s", _errstr(e))
-        content = {
-            **coerced,
-            "ok": False,
-            "warn": "response_model_validation_failed",
-        }
-        if _is_debug_mode():
-            content["warn_detail"] = _errstr(e)
-        _publish_event(
-            "decide.completed",
-            {"ok": False, "warn": "response_model_validation_failed", "request_id": coerced.get("request_id")},
-        )
-        return JSONResponse(
-            status_code=200,
-            content=content,
-        )
-
-
-@app.post(
-    "/v1/replay/{decision_id}",
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit), Depends(verify_signature)],
+from veritas_os.api.routes_memory import router as _memory_router  # noqa: E402
+from veritas_os.api.routes_memory import (  # noqa: E402,F401 -- backward compat
+    _store_put,
+    _store_get,
+    _store_search,
+    _validate_memory_kinds,
+    memory_put,
+    memory_search,
+    memory_get,
+    memory_erase,
 )
-async def replay_endpoint(decision_id: str):
-    """Replay one persisted decision and write replay report JSON."""
-    try:
-        result = await run_replay(decision_id=decision_id)
-    except ValueError:
-        return JSONResponse(
-            status_code=404,
-            content={"ok": False, "decision_id": decision_id, "error": "decision_not_found"},
-        )
-    except Exception as e:
-        logger.error("replay endpoint failed: %s", _errstr(e))
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "decision_id": decision_id, "error": "replay_failed"},
-        )
-
-    return {
-        "ok": True,
-        "decision_id": result.decision_id,
-        "replay_path": result.replay_path,
-        "match": result.match,
-        "diff_summary": result.diff_summary,
-        "replay_time_ms": result.replay_time_ms,
-    }
-
-
-@app.post(
-    "/v1/decision/replay/{decision_id}",
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)],
+from veritas_os.api.routes_trust import router as _trust_router  # noqa: E402
+from veritas_os.api.routes_trust import (  # noqa: E402,F401 -- backward compat
+    _parse_risk_from_trust_entry,
+    _prov_actor_for_entry,
+    trust_logs,
+    trust_log_by_request,
+    trust_feedback,
+    trustlog_verify,
+    trustlog_export,
+    trust_prov_export,
+    trust_log_stats,
 )
-async def replay_decision_endpoint(decision_id: str, request: Request):
-    """Replay a persisted decision deterministically and return diff report."""
-    p = get_decision_pipeline()
-    if p is None or not hasattr(p, "replay_decision"):
-        return JSONResponse(
-            status_code=503,
-            content={
-                "match": False,
-                "diff": {"error": DECIDE_GENERIC_ERROR},
-                "replay_time_ms": 0,
-            },
-        )
-
-    mock_external_apis = True
-    try:
-        qv = request.query_params.get("mock_external_apis")
-        if qv is not None:
-            mock_external_apis = str(qv).strip().lower() not in {"0", "false", "no", "off"}
-    except Exception:
-        mock_external_apis = True
-
-    try:
-        result = await p.replay_decision(
-            decision_id=decision_id,
-            mock_external_apis=mock_external_apis,
-        )
-    except Exception as e:
-        logger.error("decision replay failed: %s", _errstr(e))
-        return JSONResponse(
-            status_code=500,
-            content={
-                "match": False,
-                "diff": {"error": "replay_failed"},
-                "replay_time_ms": 0,
-            },
-        )
-    return result
-
-
-# ==============================
-# FUJI quick validate
-# ==============================
-
-def _call_fuji(fc: Any, action: str, context: dict) -> dict:
-    """
-    validate_action / validate の微妙なシグネチャ差を吸収する。
-    """
-    if hasattr(fc, "validate_action"):
-        fn = fc.validate_action
-        try:
-            return fn(action=action, context=context)
-        except TypeError:
-            return fn(action, context)
-    if hasattr(fc, "validate"):
-        fn = fc.validate
-        try:
-            return fn(action=action, context=context)
-        except TypeError:
-            try:
-                return fn(action, context)  # type: ignore
-            except TypeError:
-                return fn(action)  # type: ignore
-    raise RuntimeError("fuji_core has neither validate_action nor validate")
-
-
-@app.post(
-    "/v1/fuji/validate",
-    response_model=FujiDecision,
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)],
+from veritas_os.api.routes_governance import router as _governance_router  # noqa: E402
+from veritas_os.api.routes_governance import (  # noqa: E402,F401 -- backward compat
+    require_governance_access,
+    _governance_rbac_enabled,
+    _resolve_governance_allowed_roles,
+    _emit_governance_change_alert,
+    governance_get,
+    governance_put,
+    governance_value_drift,
+    governance_policy_history,
 )
-def fuji_validate(payload: dict):
-    if not _is_direct_fuji_api_enabled():
-        return JSONResponse(
-            status_code=403,
-            content={
-                "detail": (
-                    "direct_fuji_api_disabled: use /v1/decide pipeline or set "
-                    "VERITAS_ENABLE_DIRECT_FUJI_API=1"
-                )
-            },
-        )
-
-    fc = get_fuji_core()
-    if fc is None:
-        logger.warning("fuji_validate: fuji_core unavailable: %s", _fuji_state.err)
-        # Return 503 as JSONResponse to ensure proper format
-        return JSONResponse(
-            status_code=503,
-            content={"detail": "fuji_core unavailable"}
-        )
-
-    # ★ セキュリティ修正: action フィールドのサイズ制限（DoS対策）
-    action = str(payload.get("action", "") or "")[:10_000]
-    context = payload.get("context") or {}
-
-    try:
-        result = _call_fuji(fc, action, context)
-    except RuntimeError as e:
-        # Check if this is the "neither validate_action nor validate" error
-        # Note: Using string matching is fragile but matches test expectations
-        err_msg = str(e)
-        if "neither validate_action nor validate" in err_msg:
-            # This specific error should return 500 as expected by test_fuji_validate_no_impl_raises_500
-            logger.error("fuji_validate: %s", err_msg)
-            return JSONResponse(
-                status_code=500,
-                content={"detail": err_msg}
-            )
-        # Other RuntimeErrors: return 200 with error structure
-        # ★ セキュリティ修正: 内部エラー詳細をレスポンスに含めない
-        logger.error("fuji_validate RuntimeError: %s", err_msg)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "status": "error",
-                "reasons": ["Validation failed"],
-                "violations": []
-            }
-        )
-    except Exception as e:
-        # All other exceptions: return 200 with error structure
-        # ★ セキュリティ修正: 内部エラー詳細をレスポンスに含めない
-        logger.error("fuji_validate error: %s", _errstr(e))
-        return JSONResponse(
-            status_code=200,
-            content={
-                "status": "error",
-                "reasons": ["Validation failed"],
-                "violations": []
-            }
-        )
-
-    coerced = _coerce_fuji_payload(result, action=action)
-    if str(coerced.get("status", "")).lower() in {"reject", "rejected", DECISION_REJECTED}:
-        _publish_event(
-            "fuji.rejected",
-            {"action": action, "status": coerced.get("status"), "reasons": coerced.get("reasons", [])},
-        )
-    try:
-        return FujiDecision.model_validate(coerced)
-    except Exception as e:
-        # 最後の保険：落ちるより返す
-        # ★ セキュリティ修正: 内部エラー詳細はデバッグモード時のみ公開
-        logger.error("FujiDecision validation failed: %s", _errstr(e))
-        content = {
-            **coerced,
-            "warn": "response_model_validation_failed",
-        }
-        if _is_debug_mode():
-            content["warn_detail"] = _errstr(e)
-        _publish_event(
-            "decide.completed",
-            {"ok": False, "warn": "response_model_validation_failed", "request_id": coerced.get("request_id")},
-        )
-        return JSONResponse(
-            status_code=200,
-            content=content,
-        )
-
-
-# ==============================
-# Memory API
-# ==============================
-
-def _store_put(store: Any, user_id: str, key: str, value: dict) -> None:
-    if not hasattr(store, "put"):
-        raise RuntimeError("store.put not found")
-    fn = store.put
-    try:
-        fn(user_id, key=key, value=value)
-        return
-    except TypeError:
-        pass
-    try:
-        fn(user_id, key, value)
-        return
-    except TypeError:
-        fn(key, value)  # type: ignore
-
-
-def _store_get(store: Any, user_id: str, key: str) -> Any:
-    if not hasattr(store, "get"):
-        raise RuntimeError("store.get not found")
-    fn = store.get
-    try:
-        return fn(user_id, key=key)
-    except TypeError:
-        try:
-            return fn(user_id, key)
-        except TypeError:
-            return fn(key)  # type: ignore
-
-
-def _store_search(store: Any, *, query: str, k: int, kinds: Any, min_sim: float, user_id: Optional[str]) -> Any:
-    if not hasattr(store, "search"):
-        raise RuntimeError("store.search not found")
-    fn = store.search
-    # prefer richest signature
-    try:
-        return fn(query=query, k=k, kinds=kinds, min_sim=min_sim, user_id=user_id)
-    except TypeError:
-        pass
-    try:
-        return fn(query=query, k=k, kinds=kinds, min_sim=min_sim)
-    except TypeError:
-        pass
-    try:
-        return fn(query=query, k=k)
-    except TypeError:
-        return fn(query)
-
-
-def _validate_memory_kinds(kinds: Any) -> Tuple[Optional[list[str]], Optional[str]]:
-    """Validate and normalize memory search kinds against allow-list.
-
-    Returns:
-        A tuple of (validated_kinds, error_message).
-        - validated_kinds is None when no filter is requested.
-        - error_message is set when input should be rejected.
-    """
-    if kinds is None:
-        return None, None
-
-    raw_kinds: list[Any]
-    if isinstance(kinds, str):
-        raw_kinds = [kinds]
-    elif isinstance(kinds, list):
-        raw_kinds = kinds
-    else:
-        return None, "kinds must be a string or list of strings"
-
-    normalized_kinds: list[str] = []
-    invalid_kinds: list[str] = []
-
-    for raw_kind in raw_kinds:
-        if not isinstance(raw_kind, str):
-            return None, "kinds must be a string or list of strings"
-        normalized_kind = raw_kind.strip().lower()
-        if normalized_kind not in VALID_MEMORY_KINDS:
-            invalid_kinds.append(normalized_kind)
-            continue
-        if normalized_kind not in normalized_kinds:
-            normalized_kinds.append(normalized_kind)
-
-    if invalid_kinds:
-        return None, f"invalid kinds: {invalid_kinds}"
-
-    return normalized_kinds, None
-
-
-@app.post("/v1/memory/put", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def memory_put(body: MemoryPutRequest, x_api_key: Optional[str] = Header(default=None, alias="X-API-Key")) -> Dict[str, Any]:
-    store = get_memory_store()
-    if store is None:
-        logger.warning("memory_put: memory store unavailable: %s", _memory_store_state.err)
-        return {"ok": False, "error": "memory store unavailable"}
-
-    try:
-        user_id = _resolve_memory_user_id(body.user_id, x_api_key)
-        key = body.key or f"memory_{datetime.now(timezone.utc).strftime('%Y%m%d_%H%M%S')}"
-        value = body.value or {}
-
-        text = body.text.strip()
-        tags = body.tags
-        meta = body.meta or {}
-
-        retention_class = str(
-            body.retention_class or meta.get("retention_class") or "standard"
-        ).strip().lower()
-        expires_at = body.expires_at if body.expires_at is not None else meta.get("expires_at")
-        legal_hold = body.legal_hold or bool(meta.get("legal_hold", False))
-
-        legacy_saved = False
-        if value:
-            try:
-                _store_put(store, user_id, key, value)
-                legacy_saved = True
-            except Exception as e:
-                logger.warning("[MemoryOS][legacy] Error: %s", e)
-
-        kind = body.kind
-        if kind not in VALID_MEMORY_KINDS:
-            kind = "semantic"
-
-        new_id = None
-        if text:
-            try:
-                text_clean = redact(text)
-                meta_for_store = dict(meta)
-                meta_for_store.setdefault("user_id", user_id)
-                meta_for_store.setdefault("kind", kind)
-                meta_for_store["retention_class"] = retention_class
-                meta_for_store["expires_at"] = expires_at
-                meta_for_store["legal_hold"] = legal_hold
-
-                if hasattr(store, "put"):
-                    vector_item = {
-                        "text": text_clean,
-                        "tags": tags,
-                        "meta": meta_for_store,
-                    }
-                    try:
-                        new_id = store.put(kind, vector_item)
-                    except TypeError:
-                        if hasattr(store, "put_episode"):
-                            new_id = store.put_episode(
-                                text=text_clean,
-                                tags=tags,
-                                meta=meta_for_store,
-                            )
-                        else:
-                            raise
-                elif hasattr(store, "put_episode"):
-                    new_id = store.put_episode(
-                        text=text_clean,
-                        tags=tags,
-                        meta=meta_for_store,
-                    )
-                else:
-                    episode_key = f"episode_{int(time.time())}"
-                    _store_put(
-                        store,
-                        user_id,
-                        episode_key,
-                        {"text": text_clean, "tags": tags, "meta": meta_for_store},
-                    )
-                    new_id = episode_key
-            except Exception as e:
-                logger.warning("[MemoryOS][vector] Error: %s", e)
-
-        return {
-            "ok": True,
-            "legacy": {"saved": legacy_saved, "key": key if legacy_saved else None},
-            "vector": {
-                "saved": bool(new_id),
-                "id": new_id,
-                "kind": kind if new_id else None,
-                "tags": tags if new_id else None,
-            },
-            "size": len(str(value)) if value else len(text),
-            "lifecycle": {
-                "retention_class": retention_class,
-                "expires_at": expires_at,
-                "legal_hold": legal_hold,
-            },
-        }
-
-    except Exception as e:
-        logger.error("[MemoryOS] Error: %s", e)
-        return {"ok": False, "error": "memory operation failed"}
-
-
-@app.post("/v1/memory/search", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def memory_search(payload: MemorySearchRequest, x_api_key: Optional[str] = Header(default=None, alias="X-API-Key")) -> Dict[str, Any]:
-    store = get_memory_store()
-    if store is None:
-        logger.warning("memory_search: memory store unavailable: %s", _memory_store_state.err)
-        return {"ok": False, "error": "memory store unavailable", "hits": [], "count": 0}
-
-    try:
-        q = payload.query
-        kinds = payload.kinds
-        k = payload.k
-        min_sim = payload.min_sim
-        user_id = _resolve_memory_user_id(payload.user_id, x_api_key)
-
-        # Validate kinds against allow-list to prevent backend-specific injection.
-        validated_kinds, kinds_error = _validate_memory_kinds(kinds)
-        if kinds_error:
-            return {
-                "ok": False,
-                "error": kinds_error,
-                "hits": [],
-                "count": 0,
-            }
-
-        raw_hits = _store_search(
-            store,
-            query=q,
-            k=k,
-            kinds=validated_kinds,
-            min_sim=min_sim,
-            user_id=user_id,
-        )
-
-        # ★ Bug fix: MemoryStore.search() returns Dict[str, List[Dict]], not a flat list.
-        # Flatten the dict-of-lists into a single list before filtering.
-        if isinstance(raw_hits, dict):
-            flat_hits: list = []
-            for kind_hits in raw_hits.values():
-                if isinstance(kind_hits, list):
-                    flat_hits.extend(kind_hits)
-            raw_hits = flat_hits
-        elif not isinstance(raw_hits, list):
-            raw_hits = list(raw_hits) if raw_hits else []
-
-        norm_hits = []
-        for h in (raw_hits or []):
-            if isinstance(h, dict):
-                meta = h.get("meta") or {}
-                # ★ セキュリティ修正: 常に user_id でフィルタリング
-                if meta.get("user_id") == user_id:
-                    norm_hits.append(h)
-            else:
-                # ★ セキュリティ修正: non-dict ヒット（生ID文字列等）は user_id を
-                # 検証できないため、情報漏洩防止のためスキップする
-                continue
-
-        return {"ok": True, "hits": norm_hits, "count": len(norm_hits)}
-
-    except Exception as e:
-        logger.error("[MemoryOS][search] Error: %s", e)
-        return {"ok": False, "error": "memory search failed", "hits": [], "count": 0}
-
-
-@app.post("/v1/memory/get", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def memory_get(body: MemoryGetRequest, x_api_key: Optional[str] = Header(default=None, alias="X-API-Key")) -> Dict[str, Any]:
-    store = get_memory_store()
-    if store is None:
-        # Do not expose internal error details to clients
-        return {"ok": False, "error": "memory store unavailable", "value": None}
-
-    try:
-        uid = _resolve_memory_user_id(body.user_id, x_api_key)
-        key = body.key
-        value = _store_get(store, uid, key)
-        return {"ok": True, "value": value}
-    except Exception as e:
-        logger.error("memory_get failed: %s", e)
-        return {"ok": False, "error": "memory retrieval failed", "value": None}
-
-
-@app.post(
-    "/v1/memory/erase",
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)],
+from veritas_os.api.routes_system import router as _system_router  # noqa: E402
+from veritas_os.api.routes_system import public_router as _system_public_router  # noqa: E402
+from veritas_os.api.routes_system import events_router as _events_router  # noqa: E402
+from veritas_os.api.routes_system import (  # noqa: E402,F401 -- backward compat
+    ComplianceConfigBody,
+    SystemHaltRequest,
+    SystemResumeRequest,
+    _collect_recent_decide_files,
+    root,
+    health,
+    status,
+    metrics,
+    events,
+    trustlog_ws,
+    compliance_get_config,
+    compliance_put_config,
+    report_eu_ai_act,
+    report_governance,
+    system_halt,
+    system_resume,
+    system_halt_status,
+    compliance_deployment_readiness,
 )
-def memory_erase(body: MemoryEraseRequest, x_api_key: Optional[str] = Header(default=None, alias="X-API-Key")) -> Dict[str, Any]:
-    """Erase a tenant's memories with legal-hold protection and audit log."""
-    store = get_memory_store()
-    if store is None:
-        return {"ok": False, "error": "memory store unavailable"}
 
-    try:
-        user_id = _resolve_memory_user_id(body.user_id, x_api_key)
-        reason = body.reason.strip()
-        actor = body.actor.strip()
-
-        if hasattr(store, "erase_user"):
-            report = store.erase_user(user_id=user_id, reason=reason, actor=actor)
-            report.setdefault("ok", True)
-            return report
-
-        return {
-            "ok": False,
-            "error": "erase operation unsupported by active memory backend",
-        }
-    except Exception as e:
-        logger.error("memory_erase failed: %s", e)
-        return {"ok": False, "error": "memory erase failed"}
-
-
-# ==============================
-# metrics for Doctor
-# ==============================
-
-
-def _collect_recent_decide_files(shadow_dir: Path, limit: int) -> Tuple[list[Path], int]:
-    """Return latest decide files while keeping memory usage bounded."""
-    total_count = 0
-    min_heap: list[tuple[str, Path]] = []
-    for file_path in shadow_dir.glob("decide_*.json"):
-        total_count += 1
-        item = (file_path.name, file_path)
-        if len(min_heap) < limit:
-            heapq.heappush(min_heap, item)
-            continue
-        if item[0] > min_heap[0][0]:
-            heapq.heapreplace(min_heap, item)
-
-    newest = sorted((path for _, path in min_heap), key=lambda file_path: file_path.name)
-    return newest, total_count
-
-
-@app.get("/v1/metrics", dependencies=[Depends(require_api_key)])
-def metrics(decide_file_limit: int = Query(default=500, ge=1, le=5000)):
-    shadow_dir = _effective_shadow_dir()
-    _, _, log_jsonl = _effective_log_paths()
-
-    files, total_decide_files = _collect_recent_decide_files(shadow_dir, decide_file_limit)
-    last_at = None
-    if files:
-        try:
-            with open(files[-1], encoding="utf-8") as f:
-                last_at = json.load(f).get("created_at")
-        except Exception:
-            logger.debug("failed to read last decide file: %s", files[-1], exc_info=True)
-
-    lines = 0
-    try:
-        if log_jsonl.exists():
-            with open(log_jsonl, encoding="utf-8") as f:
-                for _ in f:
-                    lines += 1
-    except Exception as e:
-        logger.warning("read trust_log.jsonl failed: %s", _errstr(e))
-
-    result = {
-        "decide_files": total_decide_files,
-        "decide_files_returned": len(files),
-        "decide_files_truncated": total_decide_files > len(files),
-        "trust_jsonl_lines": lines,
-        "last_decide_at": last_at,
-        "server_time": utc_now_iso_z(),
-        "pipeline_ok": get_decision_pipeline() is not None,
-        "auth_store_mode": (os.getenv("VERITAS_AUTH_SECURITY_STORE") or "memory").strip().lower() or "memory",
-        "auth_store_failure_mode": _auth_store_failure_mode(),
-        "auth_reject_reasons": _snapshot_auth_reject_reason_metrics(),
-    }
-    # ★ M-13 修正: 内部エラー詳細はデバッグモード時のみ公開
-    if _is_debug_mode():
-        result["pipeline_error"] = _pipeline_state.err
-        result["cfg_error"] = _cfg_state.err
-    else:
-        result["pipeline_error"] = bool(_pipeline_state.err)
-        result["cfg_error"] = bool(_cfg_state.err)
-    return result
-
-
-@app.get("/v1/events", dependencies=[Depends(require_api_key_header_or_query)])
-async def events(request: Request, heartbeat_sec: int = Query(default=15, ge=5, le=60)):
-    """Server-Sent Events stream for near-real-time UI updates."""
-    subscriber = _event_hub.register()
-
-    async def _stream():
-        # ★ 修正: blocking な subscriber.get() を asyncio.to_thread 経由で実行し、
-        # イベントループのスレッドプールを枯渇させないようにする。
-        try:
-            yield ": connected\n\n"
-            while True:
-                if await request.is_disconnected():
-                    break
-                try:
-                    item = await asyncio.to_thread(subscriber.get, timeout=heartbeat_sec)
-                    yield _format_sse_message(item)
-                except queue.Empty:
-                    yield f": heartbeat {utc_now_iso_z()}\n\n"
-        finally:
-            _event_hub.unregister(subscriber)
-
-    headers = {
-        "Cache-Control": "no-cache",
-        "Connection": "keep-alive",
-        "X-Accel-Buffering": "no",
-    }
-    return StreamingResponse(_stream(), media_type="text/event-stream", headers=headers)
-
-
-@app.websocket("/v1/ws/trustlog")
-async def trustlog_ws(websocket: WebSocket):
-    """WebSocket stream for Debate/Critique trust logs.
-
-    Art. 12: enables real-time record visibility for technical traceability.
-    Security warning: browser WebSocket authentication uses query API key.
-    """
-    if not _authenticate_websocket_api_key(websocket):
-        await websocket.close(code=1008, reason="invalid_api_key")
-        return
-
-    await websocket.accept()
-    subscriber = _event_hub.register()
-    allowed_types = {
-        "trustlog.debate",
-        "trustlog.critique",
-        "trustlog.appended",
-        "compliance.pending_review",
-    }
-    try:
-        while True:
-            try:
-                item = await asyncio.to_thread(subscriber.get, timeout=15)
-            except queue.Empty:
-                await websocket.send_json({"type": "heartbeat", "ts": utc_now_iso_z()})
-                continue
-            if item.get("type") in allowed_types:
-                await websocket.send_json(item)
-    except WebSocketDisconnect:
-        logger.debug("trustlog ws disconnected")
-    finally:
-        _event_hub.unregister(subscriber)
-
-
-@app.get(
-    "/v1/compliance/config",
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)],
-)
-def compliance_get_config() -> Dict[str, Any]:
-    """Return runtime compliance config for UI toggle."""
-    return {"ok": True, "config": get_runtime_config()}
-
-
-@app.put(
-    "/v1/compliance/config",
-    dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)],
-)
-def compliance_put_config(body: ComplianceConfigBody) -> Dict[str, Any]:
-    """Update runtime compliance config.
-
-    Art. 9/14 controls are toggled here from the frontend switch.
-    """
-    updated = update_runtime_config(
-        eu_ai_act_mode=body.eu_ai_act_mode,
-        safety_threshold=body.safety_threshold,
-    )
-    _publish_event("compliance.config.updated", {"config": updated})
-    return {"ok": True, "config": updated}
-
-
-# ==============================
-# Trust Feedback
-# ==============================
-
-@app.get("/v1/trust/logs", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def trust_logs(cursor: Optional[str] = None, limit: int = 50):
-    """TrustLog をページング取得する。"""
-    return get_trust_log_page(cursor=cursor, limit=limit)
-
-
-@app.get("/v1/trust/{request_id}", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def trust_log_by_request(request_id: str):
-    """request_id 単位で TrustLog を取得する。"""
-    return get_trust_logs_by_request(request_id=request_id)
-
-
-@app.post("/v1/trust/feedback", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def trust_feedback(body: TrustFeedbackRequest):
-    """
-    人間からのフィードバックを trust_log に記録する簡易API。
-
-    Notes:
-        score は 0.0〜1.0 に正規化してから保存する。
-        数値変換に失敗した場合は既定値 0.5 を採用する。
-    """
-    vc = get_value_core()
-    if vc is None:
-        logger.warning("trust_feedback: value_core unavailable: %s", _value_core_state.err)
-        return {"status": "error", "detail": "value_core unavailable"}
-
-    try:
-        uid = str(body.user_id or "anon")[:500]
-        score = body.score
-        note = body.note
-        source = body.source
-        extra = {"api": "/v1/trust/feedback"}
-
-        if hasattr(vc, "append_trust_log"):
-            vc.append_trust_log(
-                user_id=uid,
-                score=score,
-                note=note,
-                source=source,
-                extra=extra,
-            )
-            _publish_event(
-                "trustlog.appended",
-                {"kind": "feedback", "user_id": uid, "source": source},
-            )
-            return {"status": "ok", "user_id": uid}
-
-        return {"status": "error", "detail": "value_core.append_trust_log not found"}
-
-    except Exception as e:
-        # Log the detailed error server-side, but do not expose it to the client.
-        logger.error("[Trust] feedback failed: %s", e)
-        return {"status": "error", "detail": "internal error in trust_feedback"}
-
-
-@app.get("/v1/trustlog/verify", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def trustlog_verify() -> Dict[str, Any]:
-    """Verify signed append-only TrustLog integrity and signatures."""
-    try:
-        return verify_trustlog_chain()
-    except Exception as e:
-        logger.error("[TrustLog] verify failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "trustlog verification failed"},
-        )
-
-
-@app.get("/v1/trustlog/export", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def trustlog_export() -> Dict[str, Any]:
-    """Export signed append-only TrustLog entries for external audit."""
-    try:
-        return export_signed_trustlog()
-    except Exception as e:
-        logger.error("[TrustLog] export failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "trustlog export failed"},
-        )
-
-
-@app.get("/v1/trust/{request_id}/prov", dependencies=[Depends(require_api_key), Depends(enforce_rate_limit)])
-def trust_prov_export(request_id: str) -> Dict[str, Any]:
-    """Export one decision trace as W3C PROV JSON for external audit tooling."""
-    try:
-        entries = get_trust_logs_by_request(request_id)
-        if not entries:
-            return JSONResponse(
-                status_code=404,
-                content={"ok": False, "error": "trust trace not found"},
-            )
-
-        latest = entries[-1]
-        prov = build_w3c_prov_document(
-            request_id=request_id,
-            decision_status=str(latest.get("decision_status") or latest.get("status") or "unknown"),
-            risk=_parse_risk_from_trust_entry(latest),
-            timestamp=str(latest.get("ts") or latest.get("timestamp") or utc_now_iso_z()),
-            actor=_prov_actor_for_entry(latest),
-        )
-        return {"ok": True, "request_id": request_id, "prov": prov}
-    except Exception as e:
-        logger.error("trust_prov_export failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to export PROV trace"},
-        )
-
-
-
-def _governance_rbac_enabled() -> bool:
-    """Return True when governance RBAC/ABAC guard is enabled."""
-    return (os.getenv("VERITAS_GOVERNANCE_ENFORCE_RBAC", "1").strip().lower()
-            not in {"0", "false", "no", "off"})
-
-
-def _resolve_governance_allowed_roles() -> set[str]:
-    """Resolve allowed governance roles from env or safe defaults."""
-    raw = (os.getenv("VERITAS_GOVERNANCE_ALLOWED_ROLES") or "admin,compliance_owner").strip()
-    roles = {item.strip().lower() for item in raw.split(",") if item.strip()}
-    return roles or {"admin", "compliance_owner"}
-
-
-def require_governance_access(
-    x_role: Optional[str] = Header(default=None, alias="X-Role"),
-    x_tenant_id: Optional[str] = Header(default=None, alias="X-Tenant-Id"),
-) -> bool:
-    """Enforce governance RBAC/ABAC constraints for admin endpoints.
-
-    Security warning:
-        If ``VERITAS_GOVERNANCE_ENFORCE_RBAC=0`` this guard is disabled and
-        governance APIs rely only on API-key authentication.
-    """
-    if not _governance_rbac_enabled():
-        logger.warning(
-            "SECURITY WARNING: governance RBAC/ABAC is disabled by "
-            "VERITAS_GOVERNANCE_ENFORCE_RBAC=0"
-        )
-        return True
-
-    role = (x_role or "").strip().lower()
-    if role not in _resolve_governance_allowed_roles():
-        raise HTTPException(status_code=403, detail="Insufficient governance role")
-
-    expected_tenant = (os.getenv("VERITAS_GOVERNANCE_TENANT_ID") or "").strip()
-    if expected_tenant and (x_tenant_id or "").strip() != expected_tenant:
-        raise HTTPException(status_code=403, detail="Tenant scope mismatch")
-
-    return True
-
-
-def _emit_governance_change_alert(previous: Dict[str, Any], updated: Dict[str, Any]) -> None:
-    """Emit a realtime alert event when high-impact governance settings change."""
-    critical_keys = ("fuji_rules", "risk_thresholds", "auto_stop")
-    changed = [key for key in critical_keys if previous.get(key) != updated.get(key)]
-    if not changed:
-        return
-
-    _publish_event(
-        "governance.alert",
-        {
-            "severity": "high",
-            "changed_fields": changed,
-            "updated_by": updated.get("updated_by", "api"),
-        },
-    )
-
-
-def _parse_risk_from_trust_entry(entry: Dict[str, Any]) -> float | None:
-    """Extract risk score from trust-log payload in a backward compatible way."""
-    if not isinstance(entry, dict):
-        return None
-
-    candidates = (
-        entry.get("risk"),
-        (entry.get("gate") or {}).get("risk") if isinstance(entry.get("gate"), dict) else None,
-        (entry.get("fuji") or {}).get("risk") if isinstance(entry.get("fuji"), dict) else None,
-    )
-    for item in candidates:
-        try:
-            if item is None:
-                continue
-            return float(item)
-        except (TypeError, ValueError):
-            continue
-    return None
-
-
-def _prov_actor_for_entry(entry: Dict[str, Any]) -> str:
-    """Resolve PROV agent label from trust entry metadata."""
-    for key in ("updated_by", "actor", "user_id", "request_user_id"):
-        value = entry.get(key)
-        if isinstance(value, str) and value.strip():
-            return value.strip()
-    return "veritas_api"
-
-
-# ==============================
-# Governance Policy API
-# ==============================
-
-@app.get("/v1/governance/value-drift", dependencies=[Depends(require_api_key), Depends(require_governance_access)])
-def governance_value_drift(telos_baseline: float = Query(default=0.5, ge=0.0, le=1.0)):
-    """Return ValueCore drift metrics against a Telos baseline."""
-    try:
-        result = get_value_drift(telos_baseline=telos_baseline)
-        return {"ok": True, "value_drift": result}
-    except Exception as e:
-        logger.error("governance_value_drift failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to load value drift metrics"},
-        )
-
-
-@app.get("/v1/governance/policy", dependencies=[Depends(require_api_key), Depends(require_governance_access)])
-def governance_get():
-    """Return the current governance policy."""
-    try:
-        policy = get_policy()
-        return {"ok": True, "policy": policy}
-    except Exception as e:
-        logger.error("governance_get failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to load governance policy"},
-        )
-
-
-@app.put("/v1/governance/policy", dependencies=[Depends(require_api_key), Depends(require_governance_access)])
-def governance_put(body: dict):
-    """Update the governance policy (partial merge)."""
-    try:
-        enforce_four_eyes_approval(body)
-        previous = get_policy()
-        updated = update_policy(body)
-        _publish_event(
-            "governance.updated",
-            {"updated_by": updated.get("updated_by", "api")},
-        )
-        _emit_governance_change_alert(previous=previous, updated=updated)
-        return {"ok": True, "policy": updated}
-    except PermissionError as e:
-        logger.warning("governance_put rejected: %s", e)
-        return JSONResponse(
-            status_code=403,
-            content={"ok": False, "error": "governance approval validation failed"},
-        )
-    except Exception as e:
-        logger.error("governance_put failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to update governance policy"},
-        )
-
-
-@app.get("/v1/governance/policy/history", dependencies=[Depends(require_api_key), Depends(require_governance_access)])
-def governance_policy_history(limit: int = Query(default=50, ge=1, le=500)):
-    """Return recent governance policy change history (newest first).
-
-    Each record contains the previous and new policy state along with who
-    made the change and when, providing a full compliance audit trail.
-    """
-    try:
-        records = get_policy_history(limit=limit)
-        return {"ok": True, "count": len(records), "history": records}
-    except Exception as e:
-        logger.error("governance_policy_history failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to load governance policy history"},
-        )
-
-
-@app.get("/v1/trust/stats", dependencies=[Depends(require_api_key)])
-def trust_log_stats():
-    """Return TrustLog append success/failure counters for operational monitoring.
-
-    Persistent failures indicate a broken hash chain or storage issue and
-    should trigger an alert in production environments.
-    """
-    try:
-        from veritas_os.logging.trust_log import get_trust_log_stats
-        stats = get_trust_log_stats()
-        return {"ok": True, **stats}
-    except Exception as e:
-        logger.error("trust_log_stats failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to read trust log stats"},
-        )
-
-
-@app.get("/v1/report/eu_ai_act/{decision_id}", dependencies=[Depends(require_api_key)])
-def report_eu_ai_act(decision_id: str):
-    """Generate an enterprise-ready EU AI Act compliance report."""
-    try:
-        result = generate_eu_ai_act_report(decision_id)
-        if not result.get("ok"):
-            return JSONResponse(status_code=404, content=result)
-        return result
-    except Exception as e:
-        logger.error("report_eu_ai_act failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to generate EU AI Act report"},
-        )
-
-
-@app.get("/v1/report/governance", dependencies=[Depends(require_api_key)])
-def report_governance(from_: str = Query(alias="from"), to: str = Query(alias="to")):
-    """Generate internal governance report for the requested date range."""
-    try:
-        result = generate_internal_governance_report((from_, to))
-        return result
-    except ValueError:
-        return JSONResponse(
-            status_code=400,
-            content={
-                "ok": False,
-                "error": "Invalid date format. Use ISO-8601 (e.g. 2026-01-01T00:00:00Z)",
-            },
-        )
-    except Exception as e:
-        logger.error("report_governance failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to generate governance report"},
-        )
-
-
-# ---------------------------------------------------------------------------
-# Art. 14(4): System halt / resume API endpoints
-# ---------------------------------------------------------------------------
-
-
-class SystemHaltRequest(BaseModel):
-    """Request body for POST /v1/system/halt."""
-
-    reason: str = Field(..., min_length=1, max_length=500)
-    operator: str = Field(..., min_length=1, max_length=200)
-
-
-class SystemResumeRequest(BaseModel):
-    """Request body for POST /v1/system/resume."""
-
-    operator: str = Field(..., min_length=1, max_length=200)
-    comment: str = Field(default="", max_length=500)
-
-
-@app.post("/v1/system/halt", dependencies=[Depends(require_api_key)])
-def system_halt(body: SystemHaltRequest):
-    """Halt the AI decision system (Art. 14(4) emergency stop).
-
-    When halted, all ``/v1/decide`` requests are refused until a human
-    operator explicitly resumes the system.
-    """
-    try:
-        result = SystemHaltController.halt(
-            reason=body.reason, operator=body.operator,
-        )
-        return {"ok": True, **result}
-    except Exception as e:
-        logger.error("system_halt failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to halt system"},
-        )
-
-
-@app.post("/v1/system/resume", dependencies=[Depends(require_api_key)])
-def system_resume(body: SystemResumeRequest):
-    """Resume the AI decision system after a halt (Art. 14(4)).
-
-    Only authorised operators should call this endpoint.  The resumption
-    is audit-logged with operator identity and an optional comment.
-    """
-    try:
-        result = SystemHaltController.resume(
-            operator=body.operator, comment=body.comment,
-        )
-        return {"ok": True, **result}
-    except Exception as e:
-        logger.error("system_resume failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to resume system"},
-        )
-
-
-@app.get("/v1/system/halt-status", dependencies=[Depends(require_api_key)])
-def system_halt_status():
-    """Return the current halt status of the AI decision system."""
-    try:
-        status = SystemHaltController.status()
-        return {"ok": True, **status}
-    except Exception as e:
-        logger.error("system_halt_status failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to retrieve halt status"},
-        )
-
-
-# ---------------------------------------------------------------------------
-# P1-5: Deployment readiness check
-# ---------------------------------------------------------------------------
-
-
-@app.get("/v1/compliance/deployment-readiness", dependencies=[Depends(require_api_key)])
-def compliance_deployment_readiness():
-    """Check deployment readiness for EU AI Act compliance (P1-5).
-
-    Validates model card freshness, bias assessment status, and DPA
-    checklist to determine if the system is ready for high-risk deployment.
-    """
-    try:
-        result = validate_deployment_readiness()
-        return {"ok": True, **result}
-    except Exception as e:
-        logger.error("compliance_deployment_readiness failed: %s", e)
-        return JSONResponse(
-            status_code=500,
-            content={"ok": False, "error": "Failed to check deployment readiness"},
-        )
+# Include routers with appropriate auth dependencies.
+# Auth functions are defined above this point, so no circular import occurs.
+_auth_deps = [Depends(require_api_key), Depends(enforce_rate_limit)]
+_gov_deps = [Depends(require_api_key), Depends(require_governance_access)]
+
+app.include_router(_system_public_router)  # health, status, root — no auth
+app.include_router(_decide_router, dependencies=_auth_deps)
+app.include_router(_memory_router, dependencies=_auth_deps)
+app.include_router(_trust_router, dependencies=_auth_deps)
+app.include_router(_governance_router, dependencies=_gov_deps)
+app.include_router(_system_router, dependencies=[Depends(require_api_key)])
+app.include_router(_events_router, dependencies=[Depends(require_api_key_header_or_query)])
